backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pillow==10.1.0
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.13.2
python-multipart==0.0.9
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Shapely==2.0.4
geojson==3.1.0
python-dotenv==1.0.1


